msgOut(false) 全部改成throw的方式返回

+++ app/Http/SingleActions/Backend/Merchant/Admin/MerchantAdminGroup/EditAction.php
@@ -38,37 +38,37 @@ class EditAction
     {
         $adminGroupELoq = $this->model::find($inputDatas['id']);
         if ($adminGroupELoq === null) {
-            throw new \Exception('300100');
+            throw new \Exception('200902');
         }
         if ($adminGroupELoq->is_super === 1) {
-            throw new \Exception('300101');
+            throw new \Exception('200905');
         }
+
         DB::beginTransaction();
-        try {
-            $adminGroupELoq->group_name = $inputDatas['group_name'];
-            $adminGroupELoq->save();
-
-            MerchantAdminAccessGroupsHasBackendSystemMenu::where('group_id', $inputDatas['id'])->delete();
-
-            //只提取当前登录管理员也拥有的权限
-            $role = Arr::wrap(json_decode($inputDatas['role'], true));
-            $role = array_intersect($contll->adminAccessGroupDetail, $role);
-
-            //添加AdminGroupDetails数据
-            $data = ['group_id' => $inputDatas['id']];
-            foreach ($role as $roleId) {
-                $data['menu_id'] = $roleId;
-                $groupDetailEloq = new MerchantAdminAccessGroupsHasBackendSystemMenu();
-                $groupDetailEloq->fill($data);
-                $groupDetailEloq->save();
-            }
-
-            DB::commit();
-            $msgOut = msgOut(true, $adminGroupELoq->toArray());
-        } catch (\Throwable $e) {
-            DB::rollBack();
-            $msgOut = msgOut(false, [], $e->getCode(), $e->getMessage());
+        $adminGroupELoq->group_name = $inputDatas['group_name'];
+        if (!$adminGroupELoq->save()) {
+            throw new \Exception('200904');
         }
+
+        MerchantAdminAccessGroupsHasBackendSystemMenu::where('group_id', $inputDatas['id'])->delete();
+
+        //只提取当前登录管理员也拥有的权限
+        $role = Arr::wrap(json_decode($inputDatas['role'], true));
+        $role = array_intersect($contll->adminAccessGroupDetail, $role);
+
+        //添加AdminGroupDetails数据
+        $data = ['group_id' => $inputDatas['id']];
+        foreach ($role as $roleId) {
+            $data['menu_id'] = $roleId;
+            $groupDetailEloq = new MerchantAdminAccessGroupsHasBackendSystemMenu();
+            $groupDetailEloq->fill($data);
+            if (!$groupDetailEloq->save()) {
+                throw new \Exception('200906');
+            }
+        }
+
+        DB::commit();
+        $msgOut = msgOut(true, $adminGroupELoq->toArray());
         return $msgOut;
     }
 }
